requirements.txt
numpy
opencv-python
pillow
matplotlib
pycryptodome